+++ R/Belgium.R
@@ -18,8 +18,8 @@ Belgium <- R6::R6Class("Belgium",
   public = list(
 
     # Core Attributes
-    #' @field country name of country to fetch data for
-    country = "Belgium",
+    #' @field origin name of origin to fetch data for
+    origin = "Belgium",
     #' @field supported_levels A list of supported levels.
     supported_levels = list("1", "2"),
     #' @field supported_region_names A list of region names in order of level.
@@ -98,8 +98,8 @@ Belgium <- R6::R6Class("Belgium",
       self$data$raw$main_broken <- self$data$raw$main
       self$data$raw$main <-
         self$data$raw$main_broken %>%
-        filter((REGION %in% self$codes_lookup[[1]]$level_1_region
-        | is.na(REGION))) %>%
+        filter((REGION %in% self$codes_lookup[[1]]$level_1_region |
+          is.na(REGION))) %>%
         bind_rows(fixed_lines)
     },