+++ NEWS.md
@@ -16,7 +16,8 @@ Thanks to @joseph-palmer, @RichardMN, and @kathsherratt for major contributions 
 * A quickstart has been added to the package README.
 
 ## Technical improvements
-* `get_regional_data()` and `get_national_data()` now use R6 method dispatch. This is an internal change and so should have minimal user impact for users of the `get_` functions. However, all datasets are now available to be used as R6 methods (see `get_available_datasets`) which may allow for more modular use cases.
+
+* `get_regional_data()` and `get_national_data()` now use R6 method dispatch. This is an internal change and so should have minimal user impact for users of the `get_` functions. However, all datasets are now available to be used as R6 methods (see `get_available_datasets`) which may allow for more modular use cases. These classes can also be initialised using `initialise_dataclass()` which is used internally by both `get_regional_data()` and `get_national_data()`.
 * Unit testing has been separated from data downloading which is now tested individually by data set. This allows for contributors to more easily assess the impact of their additions and also allows us to publish data status updates for each data sets (see the README: https://github.com/epiforecasts/covidregionaldata#readme).
 
 ## Deprecated functions
@@ -34,6 +35,7 @@ Thanks to @joseph-palmer, @RichardMN, and @kathsherratt for major contributions 
 * Mexico codes 'inegi_code' has been changed to 'inegi'.
 * UK Level 1 'ons_region_code' is now 'region_code'.
 * UK level 2 "ltla_code" is now "local_authority_code".
+* `get_available_datasets()` now return an origin column rather than a country column and a type column rather than a get_data_function to better reflect the types of data supported.
 
 
 # covidregionaldata 0.8.3